fix 32/64b build issues

+++ modules/src/render/sdl_video.hpp
@@ -31,8 +31,8 @@ private:
 	std::unique_ptr<SDL_Rect> displayrect;
 	int width, height;
 
-	uint32_t m_StartTime;
-	int m_NumFrames;
+	uint64_t m_StartTime;
+	int64_t m_NumFrames;
 };
 
 }
